@@ -1,6 +1,6 @@
 {
   "name": "material-ui-player",
-  "version": "0.2.6",
+  "version": "1.0.0",
   "types": "dist/index.d.ts",
   "description": "React/Material-UI Audio and Video Components",
   "main": "dist/index.js",
@@ -16,6 +16,7 @@
   "keywords": [
     "react",
     "material-ui",
+    "mui",
     "audio",
     "video"
   ],
@@ -30,32 +31,36 @@
   "author": "Alessio Lapi",
   "license": "MIT",
   "peerDependencies": {
-    "@material-ui/core": ">=4.0.0",
-    "@material-ui/icons": ">=4.0.0",
+    "@mui/icons-material": ">=5.0.0",
+    "@mui/material": ">=5.0.0",
+    "@mui/styles": ">=5.0.0",
     "react": ">=17.0.0",
     "react-dom": ">=17.0.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.14.8",
-    "@babel/plugin-transform-typescript": "^7.14.6",
-    "@babel/preset-env": "^7.14.9",
-    "@babel/preset-react": "^7.14.5",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@storybook/addon-actions": "^6.3.6",
-    "@storybook/addon-essentials": "^6.3.6",
-    "@storybook/addon-links": "^6.3.6",
-    "@storybook/react": "^6.3.6",
+    "@babel/core": "^7.16.0",
+    "@babel/plugin-transform-typescript": "^7.16.1",
+    "@babel/preset-env": "^7.16.4",
+    "@babel/preset-react": "^7.16.0",
+    "@emotion/react": "^11.7.0",
+    "@emotion/styled": "^11.6.0",
+    "@mui/icons-material": "^5.2.0",
+    "@mui/material": "^5.2.2",
+    "@mui/styles": "^5.2.2",
+    "@storybook/addon-actions": "^6.4.7",
+    "@storybook/addon-essentials": "^6.4.7",
+    "@storybook/addon-links": "^6.4.7",
+    "@storybook/react": "^6.4.7",
     "@types/react": "^17.0.37",
-    "babel-loader": "^8.2.2",
+    "@types/react-dom": "^17.0.11",
+    "babel-loader": "^8.2.3",
     "babel-plugin-transform-imports": "^2.0.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "storybook-addon-material-ui": "^0.9.0-alpha.24",
-    "ts-loader": "^9.2.5",
-    "typescript": "^4.3.5",
-    "webpack": "^5.48.0",
-    "webpack-cli": "^4.7.2"
-  },
-  "dependencies": {}
+    "storybook-addon-material-ui5": "^1.0.0",
+    "ts-loader": "^9.2.6",
+    "typescript": "^4.5.2",
+    "webpack": "^5.64.4",
+    "webpack-cli": "^4.9.1"
+  }
 }
